requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
pdf2image==1.17.0
numpy==1.25.2
python-dotenv==1.0.0
gunicorn==21.2.0
